requirements.txt
fastapi
uvicorn
mysql-connector-python
streamlit
requests
pydantic
